+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## 0.1.1 (2022-07-21)
+
+* disable dataset `tags` forwarding by default and add `--forward-dataset-tags` options
+* band descriptions are only forwarded if `--forward-band-tags` is set
+
 ## 0.1.0 (2022-07-21)
 
 * Initial release.